chore: refactor components to be reusable
