+++ src/client/v2/src/components/Navbar/Navbar.tsx
@@ -1,44 +1,15 @@
-import {
-  IconAdjustments,
-  IconCalendarStats,
-  IconGauge,
-  IconNotes,
-  IconPresentationAnalytics,
-} from '@tabler/icons-react';
 import { Code, Group, ScrollArea } from '@mantine/core';
-import { LinksGroup } from '../NavbarLinksGroup/NavbarLinksGroup';
+import { LinksGroup, LinksGroupProps } from '../NavbarLinksGroup/NavbarLinksGroup';
 import { UserButton } from '../UserButton/UserButton';
 import { Logo } from './Logo';
 import classes from './Navbar.module.css';
 
-const mockData = [
-  { label: 'Dashboard', icon: IconGauge },
-  {
-    label: 'Accounts',
-    icon: IconNotes,
-    initiallyOpened: true,
-    links: [
-      { label: 'Overview', link: '/' },
-      { label: 'Forecasts', link: '/' },
-      { label: 'Outlook', link: '/' },
-      { label: 'Real time', link: '/' },
-    ],
-  },
-  {
-    label: 'Transactions',
-    icon: IconCalendarStats,
-    links: [
-      { label: 'Upcoming releases', link: '/' },
-      { label: 'Previous releases', link: '/' },
-      { label: 'Releases schedule', link: '/' },
-    ],
-  },
-  { label: 'Analytics', icon: IconPresentationAnalytics },
-  { label: 'Settings', icon: IconAdjustments },
-];
+interface NavbarProps {
+  links: LinksGroupProps[];
+}
 
-export function Navbar() {
-  const links = mockData.map((item) => <LinksGroup {...item} key={item.label} />);
+export function Navbar({ links }: NavbarProps) {
+  const linkItems = links.map((item) => <LinksGroup {...item} key={item.label} />);
 
   return (
     <nav className={classes.navbar}>
@@ -50,7 +21,7 @@ export function Navbar() {
       </div>
 
       <ScrollArea className={classes.links}>
-        <div className={classes.linksInner}>{links}</div>
+        <div className={classes.linksInner}>{linkItems}</div>
       </ScrollArea>
 
       <div className={classes.footer}>
