+++ src/client/v2/src/components/NavbarLinksGroup/NavbarLinksGroup.story.tsx
@@ -1,9 +1,24 @@
-import { StoryWrapper } from '../storybook/StoryWrapper';
+import { IconCalendarStats } from '@tabler/icons-react';
+import { ComponentPreview } from '../storybook/ComponentPreview';
 import attributes from './attributes.json';
 import { NavbarLinksGroup } from './NavbarLinksGroup';
 
 export default { title: 'NavbarLinksGroup' };
 
+const mockdata = {
+  label: 'Releases',
+  icon: IconCalendarStats,
+  links: [
+    { label: 'Upcoming releases', link: '/' },
+    { label: 'Previous releases', link: '/' },
+    { label: 'Releases schedule', link: '/' },
+  ],
+};
+
 export function Usage() {
-  return <StoryWrapper attributes={attributes} component={NavbarLinksGroup} />;
+  return (
+    <ComponentPreview canvas={attributes.canvas} withSpacing>
+      <NavbarLinksGroup {...mockdata} />
+    </ComponentPreview>
+  );
 }
