+++ src/client/v2/src/components/NavbarLinksGroup/NavbarLinksGroup.test.tsx
@@ -1,11 +1,21 @@
+import { IconCalendarStats } from '@tabler/icons-react';
 import { axe, render } from '@test-utils';
-import attributes from './attributes.json';
 import { NavbarLinksGroup } from './NavbarLinksGroup';
 
+const mockdata = {
+  label: 'Releases',
+  icon: IconCalendarStats,
+  links: [
+    { label: 'Upcoming releases', link: '/' },
+    { label: 'Previous releases', link: '/' },
+    { label: 'Releases schedule', link: '/' },
+  ],
+};
+
 describe('NavbarLinksGroup', () => {
-  axe([<NavbarLinksGroup key="1" {...(attributes as any)} />]);
+  axe([<NavbarLinksGroup key="1" {...mockdata} />]);
 
   it('renders correctly', () => {
-    render(<NavbarLinksGroup {...(attributes as any)} />);
+    render(<NavbarLinksGroup {...mockdata} />);
   });
 });
