+++ src/client/v2/src/pages/Home.page.tsx
@@ -1,9 +1,42 @@
+import {
+  IconAdjustments,
+  IconCalendarStats,
+  IconGauge,
+  IconNotes,
+  IconPresentationAnalytics,
+} from '@tabler/icons-react';
 import { Navbar } from '@/components/Navbar/Navbar';
+
+const navLinks = [
+  { label: 'Dashboard', icon: IconGauge },
+  {
+    label: 'Accounts',
+    icon: IconNotes,
+    initiallyOpened: true,
+    links: [
+      { label: 'Overview', link: '/' },
+      { label: 'Forecasts', link: '/' },
+      { label: 'Outlook', link: '/' },
+      { label: 'Real time', link: '/' },
+    ],
+  },
+  {
+    label: 'Transactions',
+    icon: IconCalendarStats,
+    links: [
+      { label: 'Upcoming releases', link: '/' },
+      { label: 'Previous releases', link: '/' },
+      { label: 'Releases schedule', link: '/' },
+    ],
+  },
+  { label: 'Analytics', icon: IconPresentationAnalytics },
+  { label: 'Settings', icon: IconAdjustments },
+];
 
 export function HomePage() {
   return (
     <>
-      <Navbar />
+      <Navbar links={navLinks} />
     </>
   );
 }
